Add double click to open job functionality to view jobs grid.
Also, moved comparators to another class.
