+++ SurveyManager/SurveyManager.csproj
@@ -477,6 +477,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="utility\CEventArgs.cs" />
     <Compile Include="utility\CFileTypeConverter.cs" />
+    <Compile Include="utility\Comparators.cs" />
     <Compile Include="utility\CountyTypeConverter.cs" />
     <Compile Include="utility\CSVHelper.cs" />
     <Compile Include="utility\CustomControls\CPropertyGrid.cs">
